slider updates added
