+++ public/state-of-charge/index.html
@@ -23,7 +23,7 @@
 
     <div class="journey-specs">
       <div class="section">
-        <h2>State of charge</h2>
+        <h2>Initial state of charge</h2>
         <div class="range-slider">
           <label class="range-thumb" id="range-thumb" for="range"></label>
           <input type="range" min="0" max="435" step="1" id="range" />
