+++ state-of-charge/index.js
@@ -1,76 +1,24 @@
-import { createClient, createRequest, defaultExchanges, subscriptionExchange } from '@urql/core';
-import { pipe, subscribe } from 'wonka';
-import { SubscriptionClient } from 'subscriptions-transport-ws';
-import { createRoute, routeUpdate } from './queries.js';
 import { drawRoute } from './map.js';
 import * as mapboxPolyline from '@mapbox/polyline';
+import { fetchRoute } from './client';
+import { getStateOfCharge } from './slider';
 
 /**
- * Example application of how to build a route with the Chargetrip API.
- * Please have a look to Readme file in this repo for more details.
+ * Draw a route on a map.
  *
- * For the purpose of this example we use urgl - lightweights GraphQL client.
- * To establish a connection with Chargetrip GraphQL API you need to have an API key.
- * The key in this example is a public one and gives access only to a part of our extensive database.
- * You need a registered `x-client-id` to access the full database.
- * Read more about an authorisation in our documentation (https://docs.chargetrip.com/#authorisation).
- */
-const headers = {
-  'x-client-id': '5e8c22366f9c5f23ab0eff39',
-};
-
-const subscriptionClient = new SubscriptionClient('wss://api.chargetrip.io/graphql', {
-  reconnect: true,
-  connectionParams: headers,
-});
-
-const client = createClient({
-  url: 'https://api.chargetrip.io/graphql',
-  fetchOptions: {
-    method: 'POST',
-    headers,
-  },
-  exchanges: [
-    ...defaultExchanges,
-    subscriptionExchange({
-      forwardSubscription(operation) {
-        return subscriptionClient.request(operation);
-      },
-    }),
-  ],
-});
-
-/**
- * To create a route you need:
+ * The route object contains `polyline` data -  the polyline encoded route (series of coordinates as a single string).
+ * We need to decode this information first. We use Mapbox polyline tool (https://www.npmjs.com/package/@mapbox/polyline) for this.
+ * As a result of decoding we get pairs [latitude, longitude].
+ * To draw a route on a map we use Mapbox GL JS. This tool uses the format [longitude,latitude],
+ * so we have to reverse every pair.
  *
- * 1. Create a new route and receive back its ID;
- * 2. Subscribe to route updates in order to receive its details.
+ * @param data {object} route specification.
+ * @param id {string} route ID.
  */
-client
-  .mutation(createRoute)
-  .toPromise()
-  .then(response => {
-    const routeId = response.data.newRoute;
-    if (!routeId) return Promise.reject('Could not retrieve Route ID. The response is not valid.');
-
-    const { unsubscribe } = pipe(
-      client.executeSubscription(createRequest(routeUpdate, { id: routeId })),
-      subscribe(result => {
-        const { status, route } = result.data.routeUpdatedById;
-
-        // you can keep listening to the route changes to update route information
-        // for this example we want to only draw the initial route
-        if (status === 'done' && route) {
-          unsubscribe();
-
-          drawRoutePolyline(route); // draw a polyline on a map
-        }
-      }),
-    );
-  })
-  .catch(error => console.log(error));
-
-/**
+fetchRoute(getStateOfCharge(), routeData => {
+  drawRoutePolyline(routeData);
+});
+/*
  * Draw a route on a map.
  *
  * Route object contains `polyline` data -  the polyline encoded route (series of coordinates as a single string).
@@ -81,21 +29,9 @@ client
  *
  * @param data {object} route specification
  */
-const drawRoutePolyline = data => {
+export const drawRoutePolyline = data => {
   const decodedData = mapboxPolyline.decode(data.polyline);
   const reversed = decodedData.map(item => item.reverse());
 
-  const rangeSlider = document.getElementById('range');
-  const rangeThumb = document.getElementById('range-thumb');
-  rangeSlider.value = data.legs[0].rangeStart / 1000;
-  console.log(data.legs[0].rangeStart / 1000);
   drawRoute(reversed, data.legs);
-};
-
-const updateRangeSliderValue = () => {
-  const percent = (rangeSlider.value - 1000) / 9000;
-  const newPosition = percent * rangeSlider.offsetWidth - percent * 60;
-
-  rangeThumb.innerHTML = `${rangeSlider.value / 1000} km`;
-  rangeThumb.style.left = `calc((${newPosition}px))`;
 };
